+++ resources/lang/nl/validation.php
@@ -91,7 +91,7 @@ return [
 
     'date_time_iso_8601' => 'Het attribuut :attribute heeft geen geldig ISO 8601 datum/tijd formaat.',
 
-    'disallowed_parameter' => 'Parameter :name is niet toegestaan.',
+    'parameter_not_supported' => 'Parameter :name is niet toegestaan.',
 
     'has_one' => 'Het veld :attribute moet een naar-één relatie zijn die :types resources bevat.',