+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## [1.0.0-alpha.5] - 2021-03-12
+
+### Added
+
+- New `Rule::includePathsForPolymorph()` method. This returns an include path rule that is configured to allow all the
+  include paths for every inverse schema that a polymorphic relation has.
+
+### Fixed
+
+- The correct message is now set as the error detail when a query parameter is not supported. Previously the translation
+  key was displayed instead of the translated message.
+
 ## [1.0.0-alpha.4] - 2021-02-27
 
 ### Added